+++ website/app.py
@@ -1,76 +1,168 @@
 import os
-import pickle
-from glob import iglob
+from pathlib import Path
 
-import rdkit.Chem.AllChem as rdkit
-from flask import Blueprint, Flask, current_app, redirect, render_template, request
+from dash import Dash, dcc, html
+import dash_bootstrap_components as dbc
+from dash.dependencies import Input, Output, State, MATCH
+from dash.exceptions import PreventUpdate
+from flask import Flask
 
-
-def bb_count_fingerprint(mols, bits, radius):
-    """ """
-
-    full_fp = []
-    for mol in mols:
-        rdkit.GetSSSR(mol)
-        mol.UpdatePropertyCache(strict=False)
-        info = {}
-        fp = rdkit.GetMorganFingerprintAsBitVect(mol, radius, bits, bitInfo=info)
-        fp = list(fp)
-        for bit, activators in info.items():
-            fp[bit] = len(activators)
-        full_fp.extend(fp)
-    full_fp.append(1)
-    return full_fp
-
-
-def fingerprint():
-    bb = rdkit.AddHs(rdkit.MolFromSmiles(request.form["bb"]))
-    lk = rdkit.AddHs(rdkit.MolFromSmiles(request.form["lk"]))
-    return [bb_count_fingerprint((bb, lk), 512, 8)]
-
-
-def load_models():
-    models = {}
-    for model_path in iglob("models/*.pkl"):
-        name, _ = os.path.splitext(os.path.basename(model_path))
-        with open(model_path, "rb") as f:
-            models[name] = pickle.load(f)
-    return models
-
-
-app = Flask(__name__, instance_relative_config=True)
-app.models = load_models()
-
-default_bp = Blueprint("default", __name__, template_folder="templates")
-
-
-@default_bp.route("/")
-def root():
-    return render_template("index.html")
-
-
-@default_bp.route("/predict/<model_name>", methods=["POST"])
-def predict(model_name):
-    ans = current_app.models[model_name].predict(fingerprint())
-    return str(ans[0])
-
+from model import predict
 
 url_prefix = os.getenv("URL_PREFIX", "")
-if url_prefix:
-    static_url_path = url_prefix + "/static"
-else:
-    static_url_path = "/static"
 
-app.static_url_path = static_url_path
-app.register_blueprint(default_bp, url_prefix=url_prefix)
+server = Flask(__name__)
+app = Dash(
+    __name__,
+    external_stylesheets=[dbc.themes.FLATLY],
+    server=server,
+    requests_pathname_prefix=url_prefix + "/",
+    routes_pathname_prefix=url_prefix + "/",
+)
 
-if url_prefix:
-    # if url_prefix assume running in production
-    from werkzeug.middleware.proxy_fix import ProxyFix
 
-    app = ProxyFix(app)  # enable choice of http(s) based on request headers
+navbar = dbc.NavbarSimple(
+    children=[
+        dbc.NavItem(dbc.NavLink("Is My Cage Porous?", href="#")),
+        dbc.DropdownMenu(
+            children=[
+                dbc.DropdownMenuItem("Page 2", href="#"),
+                dbc.DropdownMenuItem("Page 3", href="#"),
+            ],
+            nav=True,
+            in_navbar=True,
+            label="More",
+        ),
+    ],
+    brand="SupraShare",
+    brand_href="#",
+    color="primary",
+    dark=True,
+)
+
+with open(Path(__file__).parent / "static" / "index.md", "r") as f:
+    md = f.readlines()
+
+content = [
+    dcc.Markdown(md),
+    dbc.Row(
+        [
+            dbc.Col(
+                [
+                    html.Img(src=app.get_asset_url("cavity.png"), height="250px"),
+                    dbc.Label(
+                        "CC3, a shape persistent cage with cavity highlighted in red."
+                    ),
+                ]
+            ),
+            dbc.Col(
+                [
+                    html.Img(src=app.get_asset_url("collapsed.png"), height="250px"),
+                    dbc.Label("A collapsed cage lacking a central cavity."),
+                ]
+            ),
+        ]
+    ),
+    html.Br(),
+    dbc.Card(
+        id="run-card",
+        children=[
+            dbc.Row(
+                [
+                    dbc.Col(dbc.Label("Linker SMILES", width="auto")),
+                    dbc.Col(dbc.Label("Building Block SMILES", width="auto")),
+                    dbc.Col(dbc.Label("Model", width="auto")),
+                    dbc.Col(),
+                    dbc.Col(dbc.Label("Result", width="auto")),
+                ],
+            ),
+        ],
+    ),
+    dbc.Button("Add Row", id="add-button"),
+    html.Br(),
+]
+
+app.layout = html.Div(
+    children=[
+        navbar,
+        html.Br(),
+        dbc.Container(
+            children=content,
+        ),
+    ]
+)
+
+
+@app.callback(
+    Output("run-card", "children"),
+    Input("add-button", "n_clicks"),
+    State("run-card", "children"),
+)
+def add_row(n_clicks, children):
+    if n_clicks is None:
+        n_clicks = 0
+
+    form = dbc.Form(
+        dbc.Row(
+            [
+                dbc.Col(
+                    dbc.Input(type="text", id=dict(type="lk", index=n_clicks)),
+                ),
+                dbc.Col(
+                    dbc.Input(type="text", id=dict(type="bb", index=n_clicks)),
+                ),
+                dbc.Col(
+                    [
+                        dbc.Select(
+                            id=dict(type="model-dropdown", index=n_clicks),
+                            value="amine2aldehyde3",
+                            options=[
+                                {"label": model.stem, "value": model.stem}
+                                for model in Path("models").glob("*")
+                            ],
+                        ),
+                    ]
+                ),
+                dbc.Col(
+                    [
+                        dbc.Button(
+                            "Run Model \U0001F680",
+                            id=dict(type="run-button", index=n_clicks),
+                            color="primary",
+                        )
+                    ],
+                    class_name="d-grid gap-2",
+                ),
+                dbc.Col(dbc.Spinner(dbc.Label(id=dict(type="result", index=n_clicks)))),
+            ],
+            align="center",
+        )
+    )
+    return children + [form]
+
+
+@app.callback(
+    Output({"type": "result", "index": MATCH}, "children"),
+    Input({"type": "run-button", "index": MATCH}, "n_clicks"),
+    State({"type": "model-dropdown", "index": MATCH}, "value"),
+    State({"type": "bb", "index": MATCH}, "value"),
+    State({"type": "lk", "index": MATCH}, "value"),
+)
+def run_model(n_clicks, model_name, bb, lk):
+    if n_clicks is None or bb is None or lk is None:
+        raise PreventUpdate
+
+    try:
+        result = predict(model_name, bb, lk)
+    except ValueError:
+        return dbc.Label("INVALID INPUT", color="warning")
+
+    if result == 1:
+        return dbc.Label("COLLAPSED", color="danger")
+    if result == 0:
+        return dbc.Label("SHAPE PERSISTENT", color="success")
+    return dbc.Label("MODEL ERROR", color="warning")
+
 
 if __name__ == "__main__":
-    # port = int(os.environ.get('PORT', 5000))
-    # app.run(host='0.0.0.0', port=port, debug=True)
-    app.run()
+    app.run_server(debug=True)
