--- website/static/index.css
@@ -1,103 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    z-index: 0;
-
-}
-
-html, body, form, main {
-    width: 100%;
-    height: 100%;
-}
-
-div {
-    float: left;
-    width: 100%;
-    height: 100%;
-    text-align: center;
-    vertical-align: middle;
-}
-
-#header {
-    /*background-color: orange;*/
-    height: 5%;
-    width: 100%;
-    margin-top: 1em;
-}
-
-.row_box {
-    width: 20%;
-}
-
-h1, h2 {
-    font-size: 1em;
-    text-align: center;
-}
-
-.title {
-    font-size: 3em;
-}
-
-.row {
-    height: 5%;
-}
-
-.intro {
-    width: initial;
-    height: initial;
-    display: inline-block;
-}
-
-.intro_p {
-    text-align: center;
-}
-
-.inner_fdiv {
-    width: 25%;
-}
-
-img {
-    max-width: 100%;
-}
-
-
-p {
-    font-size: 1em;
-    text-align: left;
-}
-
-input {
-    width: 100%;
-    height: 100%;
-}
-
-select {
-    width: 100%;
-    height: 100%;
-}
-
-button {
-    width: 100%;
-    height: 100%;
-}
-
-.a {
-    /*background-color: blue;*/
-}
-
-.b {
-    /*background-color: red;*/
-}
-
-.c {
-    /*background-color: green;*/
-}
-
-.d {
-    /*background-color: yellow;*/
-}
-
-.e {
-    /*background-color: gray;*/
-}
