--- website/static/index.html
@@ -1,110 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <header>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" type="text/css" href="index.css">
-    <script src="js/jquery.js" type="text/javascript"></script>
-    <script src="index.js" type="text/javascript"></script>
-
-  </header>
-  <body>
-      <div id="main">
-          <div class="intro">
-              <h1 class="title">Is my cage porous?</h1>
-              <br />
-              <p class="intro_p">By Lukas Turcani <a href="http://www.github.com/lukasturcani/cage_prediction">Source Code</a></p>
-              <p class="intro_p">Part of <a href="http://www.jelfs-group.org">www.jelfs-group.org</a></p>
-              <p class="intro_p">This site does not collect or store any information about you or your molecules.</p>
-              <br />
-              <p>
-                  This is a very simple app which uses trained
-                  machine learning models
-                  to predict if an organic cage will be shape persistent
-                  or collapsed. These models are discussed in detail in
-                  our <a href="https://doi.org/10.26434/chemrxiv.6995018">paper</a>.
-
-              </p>
-
-              <p>
-                  To use the app, place the SMILES of the cage building
-                  block and linker in the appropriate columns. The SMILES
-                  of molecules are readily available on wikipedia or they
-                  can be generated with ChemDraw (go to Edit -> Copy As
-                  -> SMILES and
-                  then paste into the column).
-                  The "model" column specifies
-                  which machine learning model should be used to make a
-                  prediction. For example, "alkyne2 alkyne3" means a model
-                  trained on cages generated with alkyne linkers and alkyne building blocks.
-                  The "aldehyde2 amine3" model is trained on cages generated
-                  with aldehyde linkers and amine building blocks, and so on.
-                  All models assume the cage has a tetrahedral,
-                  Tri<sup>4</sup>Di<sup>6</sup>, topology from a [4+6]
-                  reaction.
-                  For the most accurate predictions, the model should match
-                  your own linkers and building blocks. In cases of faulty
-                  SMILES the model will still provide an answer.
-              </p>
-              <br />
-              <p>
-                  The database holding the cages used for training is
-                  available at DOI: <a href="https://doi.org/10.14469/hpc/4618">10.14469/hpc/4618</a>.
-              </p>
-              <br />
-              <p>
-                  An example amine2 aldehyde3 shape persistent cage
-                  is CC3, which has the
-                  linker SMILES of "N[C@H]1CCCC[C@@H]1N" and building block SMILES of
-                  "O=Cc1cc(C=O)cc(C=O)c1". An example of a collapsed cage can be seen
-                  using the linker "NCCCCCCN" and building block "O=Cc1cc(C=O)cc(C=O)c1".
-              </p>
-
-              <div class="fdiv">
-                  <div class="inner_fdiv">
-                      <figure>
-                          <img src="cavity.png" />
-                          <figcaption>CC3, a shape persistent cage with cavity highlighted in red.</figcaption>
-                      </figure>
-                  </div>
-                  <div class="inner_fdiv">
-                      <figure>
-                          <img src="collapsed.png" />
-                          <figcaption>A collapsed cage lacking a central cavity.</figcaption>
-                      </figure>
-                  </div>
-              </div>
-
-          </div>
-
-
-
-
-          <div id="header">
-              <div class="row_box a"><h1>Linker SMILES</h1></div>
-              <div class="row_box b"><h1>Building block SMILES</h1></div>
-              <div class="row_box c"><h1>Model</h1></div>
-              <div class="row_box d"><h1>Prediction</h1></div>
-              <div class="row_box e"></div>
-          </div>
-
-          <div class="row" id="row1">
-              <div class="row_box"><input id="lk_input"/></div>
-              <div class="row_box"><input id="bb_input"/></div>
-              <div class="row_box"><select><option value="aldehyde2amine3">aldehyde2 amine3</option>
-                                           <option value="alkene2alkene3">alkene2 alkene3</option>
-                                           <option value="alkyne2alkyne3">alkyne2 alkyne3</option>
-                                           <option value="amine2aldehyde3">amine2 aldehyde3</option>
-                                           <option value="amine2carboxylic_acid3">amine2 carboxylic_acid3</option>
-                                           <option value="carboxylic_acid2amine3">carboxylic_acid2 amine3</option>
-                                   </select></div>
-              <div class="row_box" id="ans"></div>
-              <div class="row_box"><button class="predict_button" type="button">Predict</button></div>
-          </div>
-
-          <div class="row" id="add_row"><button class="add_row_button" type="button">Add row</button></div>
-
-
-      </div>
-  </body>
-
-</html>
